+++ src/pages/Hall/Hall.js
@@ -1,12 +1,12 @@
 import { React, useState, useEffect } from 'react'
-import Input from '../components/Input/Input';
-import Button from '../components/Button/Button';
-import Products from '../components/Products/Products';
+import Input from '../../components/Input/Input';
+import Button from '../../components/Button/Button';
+import Products from '../../components/Products/Products';
 
 import '../components/Products/products.css';
 import '../global.css';
 import '../styles/hall.css'
-import CartItem from '../components/CartItem/CartItem';
+import CartItem from '../../components/CartItem/CartItem';
 // import './styles/order.css';
 
 function Hall() {
