+++ src/pages/Register/Register.js
@@ -3,10 +3,10 @@ import { Link, useHistory } from 'react-router-dom';
 
 import '../styles/register.css';
 import logo from '../img/logo.png'
-import { registerUser } from "../auth";
-import { validation } from '../validation';
-import Input from '../components/Input';
-import Button from '../components/Button';
+import { registerUser } from "../../auth";
+import { validation } from '../../validation';
+import Input from '../../components/Input/Input';
+import Button from '../../components/Button/Button';
 import '../styles/login.css';
 import '../global.css';
