+++ src/pages/Register/register.css
